+++ docs/misc/changelog.rst
@@ -3,7 +3,7 @@
 Changelog
 ==========
 
-Release 1.7.0a10 (WIP)
+Release 1.7.0a11 (WIP)
 --------------------------
 
 Breaking Changes:
@@ -11,7 +11,7 @@ Breaking Changes:
 - Removed deprecated ``create_eval_env``, ``eval_env``, ``eval_log_path``, ``n_eval_episodes`` and ``eval_freq`` parameters,
   please use an ``EvalCallback`` instead
 - Removed deprecated ``sde_net_arch`` parameter
-- Upgraded to Stable-Baselines3 >= 1.7.0a10
+- Upgraded to Stable-Baselines3 >= 1.7.0a11
 
 New Features:
 ^^^^^^^^^^^^^
@@ -41,6 +41,7 @@ Others:
 - Removed ignores on Flake8 F401 (@ZikangXiong)
 - Upgraded GitHub CI/setup-python to v4 and checkout to v3
 - Set tensors construction directly on the device
+- Standardized the use of ``from gym import spaces``
 
 Release 1.6.2 (2022-10-10)
 --------------------------